Authorization and  Policy

+++ app/Http/Controllers/PostController.php
@@ -33,10 +33,7 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        if (!$post->ownedBy(auth()->user())) {
-            dd('no');
-        }
-
+        $this->authorize('delete', $post);
         
         $post->delete();
